+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.5.4.5b3] - 2022-10-17
 ### Added
 - Support for scenarioSeq in ScenarioDbManager
+- OptimizationEngine.create_do_model() to instantiate as CPOptimizer model
+- OptimizationEngine CPOptimizer methods to create dvar as DataFrame columns 
 
 ## [0.5.4.5b2] - 2022-10-02
 ### Changed